pyproject.toml
[tool.poetry]
name = "dbt2pdf"
version = "0.0.1"
description = "Generate a PDF document from your DBT project's documentation."
authors = ["Dribia Data Research <code@dribia.com>"]
maintainers = ["Albert Iribarne <iribarne@dribia.com>", "Dribia Data Research <code@dribia.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/dribia/dbt2pdf"
homepage = "https://github.com/dribia/dbt2pdf"
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Information Technology",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Topic :: Database",
    "Topic :: Documentation",
    "Topic :: Software Development :: Documentation",
]
keywords = ["dbt", "pdf", "docs", "documentation"]

[tool.poetry.dependencies]
python = "^3.10"
typer = "^0.12.5"
rich = "^13.8.0"
fpdf2 = "^2.7.9"
matplotlib = "^3.9.2"
pydantic = "^2.8.2"
dbt-artifacts-parser = "^0.7.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.8.0"


[tool.poetry.group.lint.dependencies]
ruff = "^0.6.2"
mypy = "^1.11.2"
types-fpdf2 = "^2.7.9.20240813"


[tool.poetry.group.test.dependencies]
pytest = "^8.3.2"
pytest-cov = "^5.0.0"


[tool.ruff]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
    "D",  # pydocstyle
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "C901",  # function is too complex
    "D206",  # indent with spaces, not recommended with formatter
    "W191",  # tab-indentation, not recommended with formatter
    "D203",  # no-blank-line-before-class
    "D213",  # multi-line-summary-second-line
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D"]
"dbt2pdf/cli.py" = ["B008"]

[tool.ruff.lint.isort]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
plugins = []
strict_optional = true
ignore_missing_imports = false

[[tool.mypy.overrides]]
module = [
    "dbt_artifacts_parser.parser",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "--verbose"
filterwarnings = 'error'
xfail_strict = true
testpaths = ["tests"]
markers = [
  "integration: integration tests",
]


[tool.coverage.run]
source = ["dbt2pdf"]
branch = true

[tool.coverage.report]
precision = 2
exclude_lines = [
    'pragma: no cover',
    'raise NotImplementedError',
    'if TYPE_CHECKING:',
    '@overload',
]

[tool.coverage.xml]
output = "coverage.xml"

[tool.poetry.scripts]
dbt2pdf = "dbt2pdf.cli:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
